separating query logic from csv logic
